Add files via upload
Python provides numerous libraries and tools for data preprocessing, feature selection, model training, and evaluation, such as pandas, scikit-learn, and TensorFlow. These resources make it easier to implement the sales prediction workflow and facilitate the analysis of sales data.

It's important to note that the success of sales prediction using Python depends on the quality and relevance of the dataset, appropriate feature selection, choosing the right algorithm, and continuous evaluation and refinement of the model. Additionally, incorporating domain knowledge and considering external factors that may influence sales can further improve the accuracy of the predictions.